tools: smmute: Add VFIO-based userspace driver
We add two drivers to smmute, alongside the kernel one:

VFIO: setup direct access to the device using VFIO. The userspace driver
mmaps the device frames, maps its buffer through VFIO_MAP/UNMAP, and
handles transactions by reading and writing into the frames. No MSI
support for the moment, as this is still work in progress in the kernel.

Only one process at a time can control the device. Other processes will
have to wait for the driver to release the device's group before being
able to access the device.

Signed-off-by: Jean-Philippe Brucker <jean-philippe.brucker@arm.com>