tools: smmute: Add test script
This is the script I use for regressions testing on the SMMU test engine
and SVA support. It's fairly light, so that I can get results in
reasonable time.

	# Test normal DMA
	test_smmute

	# Test SVA
	test_smmute -t sva

	# Test SVA with VFIO
	test_smmute -t vfio

That takes under an hour on the model with a normal kernel, and a couple
more hours with all debug config (lockdep etc) enabled.

Signed-off-by: Jean-Philippe Brucker <jean-philippe.brucker@arm.com>