NOMERGE: arm64: dts: fast models: Add SMMUv3 perf monitor groups
A PMU is available for the SMMUv3 model when passing option
pci.pci_smmuv3.mmu.pmu="distributed-1", with not guarantee of accuracy.
I don't know about wired interrupts, but the "distributed-1" version can
issue MSI with the same device ID as the SMMU.

NOTE: Since this isn't the default configuration of the fast model, this
commit shouldn't go upstream.

The "arm,mmu-600-pmcg" compatible string is wrong, and only there for
testing the driver. Although the model's ID registers are consistent
with MMU-600, the device is an Architecture Envelope Model, not an
MMU-600.

Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>