samples/bpf: Fix uprobe conflict in task_fd_query
Since commit 41af3cf587f4 ("tracing/uprobe: Add multi-probe per uprobe
event support") it's not possible to register uprobes and uretprobes
with the same event name anymore. Attempting to register the uretprobe
with the same name returns with EEXIST and logs the following in the
trace error:

[ 3361.401571] trace_uprobe: error: Probe type is different from existing probe
  Command: r:uprobes/test_11959 ./task_fd_query:0x2b60
           ^

Solve the conflict by appending the uprobe kind to the event alias.

Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>