selftests/bpf: Redefine BPF_TCP_CLOSE for test_sk_storage_tracing
The test_sk_storage_tracing test expects BPF_TCP_CLOSE to be defined in
vmlinux.h. This doesn't seem to be the case for clang-built Linux,
perhaps because that symbol is only used in a BUILD_BUG_ON() in the
kernel, which seems to be a NOP for clang.

BPF_TCP_CLOSE is also defined in uapi/linux/bpf.h, which cannot be
included into the test because it conflicts with vmlinux.h.

Explicitly define __BPF_TCP_CLOSE to circumvent the problem. I haven't
thought of a cleaner solution yet.

Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>