iommu/of: Keep deferring probe for virtio-iommu
The virtio-iommu driver can be built as a module (because it depends on
virtio component commonly built as modules). Currently probe deferral
for IOMMU devices stops at initcalls, which isn't good enough when
userspace is in charge of loading the driver. Force probe deferral for
virtio-iommu.

To be fixed with fw_devlink instead.

Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>