drivers: base: Allow parts of GENERIC_CPU_DEVICES to be overridden
architectures often have extra per-cpu work that needs doing
before a CPU is registered, often to determine if a CPU is
hotpluggable.

To allow the decision on when to call register_cpu() to move to common
code, wrap the call as a __weak arch_register_cpu(). This aligns with
the way x86 and ia64 register hotplug CPUs when the become present.

ACPI's acpi_processor.c has a __weak version of this symbol because
arm64 doesn't define one. The duplicate __weak definitions are
only a problem if arm64 selects GENERIC_CPU_DEVICES without defining
one. This gets fixed up in later patches.

Signed-off-by: James Morse <james.morse@arm.com>