arm64: psci: Ignore NOT_PRESENT CPUs
For CPU hot-add, arm64 presents through ACPI a CPU with all its
resources present, but marks it unavailable. When attempting to bring
that CPU up, PSCI returns NOT_PRESENT. Handle this return code
gracefully instead of printing an error.

This requires an update to the PSCI spec. As of PSCI v1.1, NOT_PRESENT
isn't a valid return value for CPU_ON.

Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>