FIXME: vhost: vhost-iommu registration for vhost-net
This needs a cleaner solution.

vhost-net could call vhost.c to register new fds, and vhost-iommu could
then query vhost.c for this fd in VHOST_IOMMU_REGISTER_ENDPOINT

These can all be modules but both vhost-net and vhost-iommu depend on
vhost, so that should work out. The fd cannot be released until
vhost-net deregisters it.

What happens when either modules are removed?

How to extend that to VFIO?

Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>