bpf: Pass location information to bpf_arch_text_poke()
The arm64 implementation of bpf_arch_text_poke() will need to
differentiate between a function entry and its body. A function entry
may start with a Branch Target Identification instruction (BTI), which
helps protect control flow. In this case the NOP instruction to be
patched follows the BTI instruction. Otherwise, the address passed to
bpf_arch_text_poke() points to the NOP to patch.

Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>