iommu/iopf: Call device driver handler on invalid faults
At the moment we silently ignore invalid I/O page faults, whereas a
similar behavior from a process would generate a SIGSEGV signal. Let
device drivers register a fault handler to deal with invalid page faults
from the accelerators, and possibly signal the process that caused it.

There is a fundamental locking issue between this fault notification and
the call to unbind(): the device driver must not take the same lock in
handler() that it holds when calling unbind(). Since unbind() may flush
the workqueue, this would deadlock. This is annoying because the fault
handler likely won't be tested properly (as in hammer it with lockdep
enabled), so the problem will be difficult to detect. I'd prefer adding
complexity into the IOMMU driver and let device drivers do what they
want, but I don't know how at the moment.

Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>