DEBUG: KVM: arm64: Add a pl011 driver for pKVM
It might be useful at some point during development to display a
message. Add a small half-baked driver for QEMU's pl011 with a printf.

What I do is run QEMU with a virtio-serial device as the main console,
and attach the pl011, that is always instantiated on the virt machine,
to the muxed stdio:

  -nodefaults
  -chardev stdio,mux=on,id=char0,signal=off
  -device virtio-serial-device
  -device virtconsole,chardev=char0
  -mon chardev=char0,mode=readline
  # pl011
  -serial chardev:char0
  # kernel params
  -append pl011.hyp_uart=0x9000000

I still need to figure out why it breaks the monitor (can't come back
from it).

Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>