EXPERIMENT: KVM: arm64: smmu-v3: Reduce critical section in map() and unmap()
Try to improve scalability by reducing the amount of work while holding
the global lock. Remove the need to hold the lock while using domains.
This is experimental and I didn't properly check all memory barriers
(I'm testing under QEMU TCG on an x86 host). Adding lockless complexity
in the hyp code is a bad idea, so this will likely go to the trash
(although atomic refs is a classic problem and some formal modelling
would help validate it, happy to work on that).

Unfortunately it does seem to improve scalability for unmap():

                         map (x)  unmap (x)

         host 1 threads    1        1
          hyp 1 threads    2.9      1.8
    hyp+optim 1 threads    3.1      1.9

         host 4 threads    1        1
          hyp 4 threads    6.5      4.9
    hyp+optim 4 threads    6        2.8 <-

    (latency relative to !pkvm host driver, lower is better)

Global locks are still needed for:
* command queue. Could be improved with the multi-queue extension or the
  optimization of the host driver.
* power state. Maybe a reference.
* host DMA donations. This one's more complicated, and likely the
  biggest cause of the performance difference between host and hyp
  drivers.

Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>