iommu: Remove iommu_ops from bus when unregistering a device
This is a temporary fix, as the IOMMU bus ops are being removed. It's
racy but that's fine.

With the SMMUv3 KVM driver, there are two possible drivers for one
device. If the KVM driver takes ownership of the SMMUv3, and then the
hypervisor setup fails, we want to cleanly unregister the SMMUv3 KVM
driver and allow the normal driver to run. Remove the IOMMU ops when
unregistering the last device.

Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>