ACPI/VIOT: Initialize the correct IOMMU fwspec
When initializing the fwspec for a PCI device, we want to insert all
possible alias RIDs of the device into that fwspec. We use
pci_for_each_dma_alias() to find those, which calls
viot_pci_dev_iommu_init(). At the moment we incorrectly use the argument
passed to that function, which is not the original device but the bridge
aliasing that device. In order to intialize the correct fwspec, keep
track of the original device in the cookie passed to
pci_for_each_dma_alias().

Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>