KVM: arm64: Add smc64 trap handling for protected guests
The guest FF-A will be using smc64 convention which has to be handled on
a separate path. Add support for trapping these instructions for
protected guests only. Forward these request to the secure OS and add a
mockup function that will do the filtering of these calls based on an
allow list.

Signed-off-by: Sebastian Ene <sebastianene@google.com>