iommu/io-pgtable: Add unmap_leaf() operation
The KVM driver needs to retrieve the physical address and size of each
page being unmapped, in order to unshare it (allow the page to be
assigned to guests). One way of doing that is by first calling
iova_to_phys() before unmap(), though that needs to be modified to als
return the mapping size (page or block). A more efficient way is to
unmap directly and return the physical address in addition to the size.

Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>