KVM: arm64: Repaint guest shared IPA buffer addresses with PA
Iterate the list of the constituents received from the guest and repaint
the buffer addresses from IPA to PA before sharing the memory with
Trustzone. Identify if the guest is trying to share more than one page
in one constituent and break it down into multipple constituents of one
page size. This is required because a contiguous region in IPA space may
not be necessary contiguous in the physical memory.

Signed-off-by: Sebastian Ene <sebastianene@google.com>