KVM: arm64: Introduce PTE management callbacks
In preparation for supporting a different reference counting mechanism for
the host, introduce a structure that will hold all the page table
management callbacks. This currently wraps the force_pte callback and
the pte_is_counted. Pass the address of this ops structure as part of the
pagetable initialization. Store the address of the ops structure in the
pagetable structure as well as in the visiting context.

Change-Id: I6d112664fd3da5d98b068edef2ca42b6c3aa56fb
Signed-off-by: Sebastian Ene <sebastianene@google.com>