KVM: arm64: Relinquish Trusty FF-A descriptors
On VM teardown the memory that has been shared from the Guest VM with
the FF-A has to be reclaimed to prevent leaks. In order to be able to do
memory reclaim, Trustzone has to relinquish the memory access first.
Implement a workaround to notify Trusty to relinquish the memory access
for the virtio channel until the FF-A destroy message is supported.

Signed-off-by: Sebastian Ene <sebastianene@google.com>