KVM: arm64: Add tracing support for the nVHE hyp
Running at EL2, the host has very close to no way to know what's
happening in the hypervisor, which is great from a security point of
view, a bit less for debugging. With the introduction of the protected
mode, this piece of code is getting more responsibilities, which we
would like to debug and profile, hence the need to trace "things" that
are happening in the hypervisor.

There's no way the hypervisor could log things directly into the host
tracing interface. So instead let's use a separated per-CPU ring-buffer
compliant with the host so the latter can decode the events.

Signed-off-by: Vincent Donnefort <vdonnefort@google.com>