KVM: arm64: nvhe: Ignore SVE hint in SMCCC function ID
When SVE is enabled, the host may set bit 16 in SMCCC function IDs, a
hint that indicates an unused SVE state. At the moment NVHE doesn't
account for this bit when inspecting the function number, and rejects
most calls. Clear the hint bit before comparing function numbers.

Version compatibility: the host's PSCI driver initially probes for a
SMCCC version number by issuing ARM_SMCCC_VERSION_FUNC_ID, which goes to
the firmware. If the firmware implements a protocol recent enough (1.3),
subsequent SMCCC calls have the hint bit set. Since earlier versions of
the protocol reserved the hint bit, clearing it is fine regardless of
the version in use.

When a new hint is added in the future, the host will enable it after
probing the SMCCC version. Provided ARM_SMCCC_CALL_HINTS is updated
along with the host support for this new hint, NVHE will take the new
hint into account. We only clear known hints and leave reserved bits as
is, because future SMCCC versions could use reserved bits as modifiers
for the function ID.

Fixes: cfa7ff959a78 ("arm64: smccc: Support SMCCC v1.3 SVE register saving hint")
Reported-by: Ben Horgan <ben.horgan@arm.com>
Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>